fixing IsColumn()
